\name{outliersHD-package}
\alias{outliersHD-package}
\docType{package}
\title{
Detection of Outliers in High Dimensional Data
}

\description{
Functions to perform maximum likelihood estimation of and random value simulation from the scaled von Mises-Fisher distribution.
For more information see Scealy and Wood (2019).
}

\details{
\tabular{ll}{
Package: \tab outliersHD\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2026-06-15 \cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>.
}

\references{
Scealy J.L. and Wood A.T.A. (2019). Scaled von Mises-Fisher distributions and regression models for paleomagnetic directional data.
Journal of the American Statistical Association, 114(528):  1547--1560.
}





